test-client/pyproject.toml
[tool.poetry]
name = "test-client"
version = "0.1.0"
description = ""
authors = ["theoremoon <theoldmoon0602@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
